fix: build errors from _embed removal and bitcode pipeline integration
- Fix const char* vs Span<const CHAR> type mismatches in dns_client.cc,
  socket_tests.h, tls_tests.h, websocket_tests.h, file_system_tests.h
- Fix DOUBLE == double ambiguity in double_tests.h
- Fix Argument(double) ambiguous constructor in string_formatter.h
- Add bitcode pipeline for standalone pic-transform on Windows:
  compile -> transform -> assemble (via CMAKE_CXX_COMPILE_OBJECT override)

Known issue: pic-transform PHI node ordering bug causes backend crash
on some translation units (chacha20.cc). Needs fix in pic-transform pass.

+++ tests/double_tests.h
@@ -498,16 +498,16 @@ private:
 		}
 
 		// Test formatting output (this also tests that the values are properly passed through varargs)
-		if (testArray[0] == 1.1 &&
-			testArray[1] == 2.2 &&
-			testArray[2] == 3.3 &&
-			testArray[3] == 4.4 &&
-			testArray[4] == 5.5 &&
-			testArray[5] == 6.6 &&
-			testArray[6] == 7.7 &&
-			testArray[7] == 8.8 &&
-			testArray[8] == 9.9 &&
-			testArray[9] == 10.1)
+		if (testArray[0] == DOUBLE(1.1) &&
+			testArray[1] == DOUBLE(2.2) &&
+			testArray[2] == DOUBLE(3.3) &&
+			testArray[3] == DOUBLE(4.4) &&
+			testArray[4] == DOUBLE(5.5) &&
+			testArray[5] == DOUBLE(6.6) &&
+			testArray[6] == DOUBLE(7.7) &&
+			testArray[7] == DOUBLE(8.8) &&
+			testArray[8] == DOUBLE(9.9) &&
+			testArray[9] == DOUBLE(10.1))
 		{
 			return true;
 		}
